Fix commands issues

+++ cmd/zarbcli/tx/send_tx.go
@@ -69,7 +69,7 @@ func SendTx() func(c *cli.Cmd) {
 			var stamp hash.Stamp
 			var sender crypto.Address
 			var receiver crypto.Address
-			var seq int
+			var seq int32
 			var amount int64
 			var fee int64
 			var auth string
@@ -126,7 +126,7 @@ func SendTx() func(c *cli.Cmd) {
 
 			//RPC
 			if seqOpt != nil {
-				seq = *seqOpt
+				seq = int32(*seqOpt)
 			} else {
 				seq, err = grpcclient.GetSequence(promptRPCEndpoint(grpcOpt), sender)
 				if err != nil {
